feat(content): add content-files-token and use it to provide injectContentFiles function
A new CONTENT_FILES_TOKEN is created that gets the raw files
of the content folder, parses them with front-matter
and provides the content, filename, and attributes.
Then, this token is injected into the injectContentFiles
function, which is a convenience method that lets users
easily use the token in their applications.
get-raw-files.ts exists, so we can mock import.meta.glob
use away when testing the factory method
of the CONTENT_FILES_TOKEN.
The new injectContentFiles method and the ContentFiles
interface are exported from the library.