test: add config to be able to use component from libs and use it in analog-app
To ensure there is compatibility of Analog being used within an Nx
workspace, the analog-app project in this repository
is updated to use a library.
This includes installing vite-tsconfig-paths and updating the analog-apps
vite.config.ts with the plugin and setting the root to the workspace root.
This allows vite to pick up the workspace tsconfig.base.json, which
includes the necessary mapping of the import url to the lib.

Closes #259