fix: remove hard coded superjson transformer in favor of trpc client options & add e2e tests
instead of hardcoding the superjson transformer into our trpc client I removed
it and in favor of allowing the default trpc clients options object to be
passed in and forwarded to said trpc client. this allows users to pass in
whatever transformers they want and also additional links. I also added e2e
tests for the trpc-app that ensures a basic insert, read, and delete
functionality works!