fix(vite-plugin-nitro): use proxy to relay requests without api prefix
there was an issue that when built for production our nitro
server would turn POST requests into GET requests and drop
the body. This issue stems from our middleware using $fetch
to relay the request with the removed /api prefix. By default,
$fetch requests are GET requests and therefore all other
requests are converted. So instead of using $fetch directly, we
use h3's proxyRequest function to take care of the conversion.
We still pass the $fetch.native as our custom fetch implementation
as it takes care of figuring out the base URL for us. To make
sure that we also test this issue I changed the trpc-app's e2e
test to run against the production output. I did the same for the
analog-app-e2e-playwright project.