fix(vite-plugin-angular): apply babel transformations for safari 15 support
in order to get analog apps working on safari versions 15.x on iOS
to work we need to apply the babel transformations that enable
compatibility correctly. this includes a fix in the supported
browsers that expect an actual name of the browser and also
the application of the babel transformation during optimization.