feat(trpc): use consistent naming for Trpc exports and use in Nx plugin
naming for trpc is inconsistent. this deprecates the inconsistently
named variables/functions and introduces variables/functions that all
follow the Trpc... naming. While I was at it, I adjusted the nx plugin
to use those new names and improved Nx plugin testing. I also moved to
a (home).page.ts file and a analog-welcome.component.ts, which will make
it easier to delete the generated files.