fix(nx-plugin): make nx plugin use 16.1
currently the nx plugin let the nx angular init generator
pick the version of angular to be installed. this meant
that when the nx angular init generator picks 16.0.0, due
to ~16.0.0 version in the nx repository, our created analog
application did not work anymore. this fix makes sure the
correct version ^16.1.0 is picked by nx tooling. unfortunately,
semver does not like ^ in its versions so we have to remove it
before finding the major version or ensuring compatibility.