fix(trpc): cache-state matching angular & invoke macro task even on error
cache state is now active as long as the cache for the http
client is active in angular 16.1. also for now we invoke the
macro task even when an error occurs so the build does not
hang infinitely. this commit fixes the dev experience and
brings back the nice prerender on the server and avoid the
initial http request. it DOES NOT fix the build and you get
the initial request unfortunately.