fix(trpc): avoid hard coding of host/port by using $fetch/fetch when available
currently you need to know the host and port of the application in order to
hard code the trpc endpoint. this is an issue if you want to pre-render a route
calling a trpc endpoint, because during pre-rendering the application does not
run on the same host/port as in dev.
To address this issue we can simply use the $fetch provided by nitro when its available
and fall back to the default fetch when it is not.
This has a couple benefits:
1. you do no longer have to know the host/port, but can simply use /api/trpc, etc. as your endpoint
2. pre-rendering now works as expected as we query the correct nitro server

Closes #468