fix(trpc): adding host and port env variables to nitro dev process & using them for trpc during dev
$fetch is smart enough to infer the host and port for our trpc client
when it is used on the server. However, $fetch is only available during
the pre-render phase and during SSR of the prod nitro server. It is not
available during development when we work with the vite server. This PR
adds the host & port from the vite server as env variables and uses them
as fallbacks. It also adjusts the Nx plugin's and trpc-app's vite config
files to make sure that the index page is server side rendered, but not
pre-rendered to avoid not displaying the latest data.