fix(vitest-angular): reset TestBed between tests
Vitest Browser Mode caches modules and does not re-execute side-effects.
Therefore, when hooks are cleared, they are not set up again so TestBed
is not reset after each test.

Vitest was fixed to allow side-effects on setup files by
not caching them (see https://github.com/vitest-dev/vitest/pull/8884),
but the behavior is not recursive so hooks that are
registered as side effects in other files such as
`@analogjs/vitest-angular/setup-testbed.ts` will not work.

The solution is to register these hooks inside the `setUsetupTestBed` function.

Closes #1994